Little adjustments
